+++ tfrecords.py
@@ -1,11 +1,9 @@
 import ntpath
 import os
-
-import numpy as np
 import tensorflow as tf
 
 from config import FLAGS
-from utils import DEPTH, FILENAME, HEIGHT, HR_IMAGE, LR_IMAGE, TFRECORD, WIDTH, do_resize, get_image, get_tfrecord_files, load_files, parse_function, save_config
+from utils import DEPTH, FILENAME, HEIGHT, HR_IMAGE, LR_IMAGE, TFRECORD, WIDTH, get_image, get_tfrecord_files, load_files, parse_function, save_config
 
 
 def _bytes_feature(value):
@@ -18,22 +16,6 @@ def _int64_feature(value):
 
 def _float_feature(value):
     return tf.train.Feature(float_list=tf.train.FloatList(value=value.flatten()))
-
-
-def _prepare_image(file, size, colored=False):
-    image = get_image(file, size, colored)
-    image = pre_process(image)
-    return image
-
-
-def _normalize(image):
-    return image / 255.
-
-
-def pre_process(images):
-    pre_processed = _normalize(np.asarray(images))
-    pre_processed = pre_processed[:, :, np.newaxis] if len(pre_processed.shape) == 2 else pre_processed
-    return pre_processed
 
 
 def create_tfrecords(config=FLAGS):
@@ -50,8 +32,8 @@ def create_tfrecords(config=FLAGS):
         print(file)
         name = ntpath.basename(file).split('.')[0]
         lowres_filename = os.path.join(config.data_dir, config.dataset, config.subset, 'Lowres', '%s.%s' % (name, config.extension))
-        hr_image = _prepare_image(file, config.image_size, config.color_channels == 3)
-        lr_image = _prepare_image(lowres_filename, 256, config.color_channels == 3)
+        hr_image = get_image(file, config.image_size, config.color_channels == 3)
+        lr_image = get_image(lowres_filename, 256, config.color_channels == 3)
 
         # Create a feature and record
         feature = {
@@ -73,7 +55,6 @@ def create_tfrecords(config=FLAGS):
 def test_tfrecords(config=FLAGS):
     assert os.path.exists(config.tfrecord_dir)
     assert os.path.exists(os.path.join(config.tfrecord_dir, config.dataset, config.subset))
-
 
     filenames = get_tfrecord_files(config)
